refactor: minor refactor and kdoc before merge. Also fixed unused dependencies and alias
