+++ app/src/main/java/no/usn/mob3000/data/repository/content/remote/NewsRepository.kt
@@ -9,7 +9,6 @@ import no.usn.mob3000.data.source.remote.content.NewsDataSource
 import no.usn.mob3000.domain.model.content.NewsData
 import no.usn.mob3000.domain.repository.content.INewsRepository
 import java.util.UUID
-
 /**
  * Repository class responsible for managing operations related to the docs table. It uses [NewsDataSource] for fetching and handling
  * database actions. Via [INewsRepository] it makes a possible communication route with the UI domain layer, without the domain layer getting accidental access
@@ -42,7 +41,11 @@ class NewsRepository(
     }
 
     /**
-     * TODO: Is clearAllNews a little overkill in the long run? I think the function should be a little more sustainable and scalable. Thank god for garbage collection, but it will take a toll on local cache after some time
+     * Fetches a list of all news to later be used for generating news-cards in the UI.
+     * It maps the fetched data to a domain model, so it can be used in the UI.
+     *
+     * @return A result containing a list of news.
+     * @throws Exception If an error occurs during the fetching process.
      */
     override suspend fun refreshNewsFromNetwork(): Result<Unit> {
         return try {
@@ -55,8 +58,6 @@ class NewsRepository(
             Result.failure(e)
         }
     }
-
-
     /**
      * Deletes a news by its ID. The ID is directly fetched by what specific card has been opened from one of the main screens
      * [NewsScreen] -> [NewsDetailsScreen]
@@ -72,7 +73,6 @@ class NewsRepository(
             Result.failure(e)
         }
     }
-
     /**
      * Updates a chosen news by its ID. The ID is directly fetched by what specific card has been opened from one of the main screens
      * [NewsScreen] -> [NewsDetailsScreen]. As long as the ewsId exist, the update instance starts. Without an actual UUID for the news
@@ -105,7 +105,6 @@ class NewsRepository(
             return Result.failure(Exception("Original news data not found"))
         }
     }
-
     /**
      * Fetches a news by its ID. Used for populating the update screen and identify what row we are working with.
      *
@@ -120,7 +119,6 @@ class NewsRepository(
             Result.failure(e)
         }
     }
-
     /**
      * Inserts a new news.The database has an auto generated UUID for new rows, but we call it here for good measure. Not passing the value seems
      * to interfere with the null exception. It displays the time of creation using [kotlinx.datetime.Clock] and tracks what user made the row.
@@ -146,7 +144,6 @@ class NewsRepository(
         )
         return newsDataSource.insertNews(newsItem)
     }
-
     /**
      * Maps a NewsDto to a NewsData. For usage in the domain layer. Might be moved if repository are made abstract.
      */
@@ -176,7 +173,6 @@ class NewsRepository(
         )
     }
 
-
     private fun NewsItemLocal.toDomainModel(): NewsData {
         return NewsData(
             newsId = this.newsId,
@@ -189,5 +185,4 @@ class NewsRepository(
             isPublished = this.isPublished
         )
     }
-
 }
