+++ app/src/main/java/no/usn/mob3000/data/source/local/content/DocsDao.kt
@@ -7,7 +7,6 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
 import no.usn.mob3000.data.model.content.local.DocsItemLocal
-
 /**
  * Data Access Object (DAO) for the Docs entity.
  *
