+++ app/src/main/java/no/usn/mob3000/domain/usecase/content/docs/DeleteDocsUseCase.kt
@@ -1,8 +1,6 @@
 package no.usn.mob3000.domain.usecase.content.docs
 
-import no.usn.mob3000.data.repository.content.remote.DocsRepository
 import no.usn.mob3000.domain.repository.content.IDocsRepository
-
 /**
  * Use case for deleting news. Functions as a bridge between the UI and Data layers.
  *
