+++ app/src/main/java/no/usn/mob3000/domain/viewmodel/content/provideDocumentationViewModel.kt
@@ -9,7 +9,13 @@ import no.usn.mob3000.domain.usecase.content.docs.DeleteDocsUseCase
 import no.usn.mob3000.domain.usecase.content.docs.FetchDocUseCase
 import no.usn.mob3000.domain.usecase.content.docs.InsertDocsUseCase
 import no.usn.mob3000.domain.usecase.content.docs.UpdateDocsUseCase
-
+/**
+ * Provides documentationViewModel with an instance. Instead of instancing everything trough the useCases, I use this support-class for each viewModel.
+ * Having interfaces does cover some of this workarounds, but for extra measure i've collected all the initializations here.
+ *
+ * @author 258030
+ * @created 2024-11-13
+ */
 fun provideDocumentationViewModel(context: Context): DocumentationViewModel {
     val docsRepositoryLocal = DocsRepositoryLocal(context)
     val docsRepository = DocsRepository(
@@ -17,7 +23,6 @@ fun provideDocumentationViewModel(context: Context): DocumentationViewModel {
         authDataSource = AuthDataSource(),
         docsDataSource = DocsDataSource()
     )
-
     return DocumentationViewModel(
         fetchDocUseCase = FetchDocUseCase(docsRepository),
         deleteDocsUseCase = DeleteDocsUseCase(docsRepository),
